// src/ui/app/root/application/models/Board/ProgramCampus.ts
export interface IBoardProgramCampus {
    programCampusId: string;
    title: string;
    description: string;
    programId: string;
    statusId: number;
}

export interface IVWProgramCampus {
    programCampusId: string;
    title: string;
    description: string;
    programId: string;
    programName: number;
    statusId: number;
}